requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
python-dotenv>=1.0.0
jupyter>=1.0.0
ipywidgets>=8.0.0
matplotlib>=3.7.0
pytest>=7.0.0
